use the styleguide pageHeader for patient dashboards
