+++ packages/esm-appointments-app/src/appointments.component.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from 'react';
-import { useTranslation } from 'react-i18next';
 import dayjs from 'dayjs';
 import AppointmentTabs from './appointments/appointment-tabs.component';
 import AppointmentsHeader from './header/appointments-header.component';
@@ -9,7 +8,6 @@ import SelectedDateContext from './hooks/selectedDateContext';
 import { omrsDateFormat } from './constants';
 
 const Appointments: React.FC = () => {
-  const { t } = useTranslation();
   const [appointmentServiceType, setAppointmentServiceType] = useState<string>('');
   const [selectedDate, setSelectedDate] = useState<string>(dayjs().startOf('day').format(omrsDateFormat));
 
@@ -30,7 +28,7 @@ const Appointments: React.FC = () => {
   return (
     <SelectedDateContext.Provider value={{ selectedDate, setSelectedDate }}>
       <AppointmentsHeader
-        title={t('home', 'Home')}
+        title={'Appointments'}
         appointmentServiceType={appointmentServiceType}
         onChange={setAppointmentServiceType}
       />
