--- packages/esm-patient-list-management-app/src/header/header.scss
@@ -1,52 +1,0 @@
-@use '@carbon/styles/scss/spacing';
-@use '@carbon/styles/scss/type';
-@import '../style.scss';
-
-.patientListHeader {
-  @include type.type-style('body-compact-02');
-  color: $text-02;
-  height: spacing.$spacing-12;
-  display: flex;
-  justify-content: space-between;
-}
-
-.leftJustifiedItems {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.rightJustifiedItems {
-  @include type.type-style('body-compact-02');
-  display: flex;
-  flex-direction: column;
-  color: $text-02;
-  padding: 1rem 0;
-  justify-content: space-between;
-}
-
-.date {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  margin-right: 1rem;
-
-  svg {
-    margin-right: 0.5rem;
-  }
-}
-
-.pageName {
-  @include type.type-style('heading-04');
-}
-
-.pageLabels {
-  p:first-of-type {
-    margin-bottom: 0.25rem;
-  }
-}
-
-.newListButton {
-  align-self: flex-end;
-  width: fit-content;
-}
