+++ packages/esm-service-queues-app/src/patient-queue-header/patient-queue-header.component.tsx
@@ -1,8 +1,7 @@
 import React, { useCallback, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Location } from '@carbon/react/icons';
 import { Dropdown } from '@carbon/react';
-import { useConfig, useSession } from '@openmrs/esm-framework';
+import { useSession, PageHeader } from '@openmrs/esm-framework';
 import PatientQueueIllustration from './patient-queue-illustration.component';
 import { useQueueLocations } from '../patient-search/hooks/useQueueLocations';
 
@@ -14,8 +13,6 @@ import {
   useSelectedQueueLocationUuid,
 } from '../helpers/helpers';
 import styles from './patient-queue-header.scss';
-import type { ConfigObject } from '../config-schema';
-
 interface PatientQueueHeaderProps {
   title?: string | JSX.Element;
   showLocationDropdown: boolean;
@@ -25,9 +22,7 @@ interface PatientQueueHeaderProps {
 const PatientQueueHeader: React.FC<PatientQueueHeaderProps> = ({ title, showLocationDropdown, actions }) => {
   const { t } = useTranslation();
   const { queueLocations, isLoading, error } = useQueueLocations();
-  const { dashboardTitle } = useConfig<ConfigObject>();
   const userSession = useSession();
-  const userLocation = userSession?.sessionLocation?.display;
   const currentQueueLocationName = useSelectedQueueLocationName();
   const currentQueueLocationUuid = useSelectedQueueLocationUuid();
 
@@ -66,19 +61,12 @@ const PatientQueueHeader: React.FC<PatientQueueHeaderProps> = ({ title, showLoca
   ]);
 
   return (
-    <div className={styles.header} data-testid="patient-queue-header">
-      <div className={styles['left-justified-items']}>
-        <PatientQueueIllustration />
-        <div className={styles['page-labels']}>
-          <p>{dashboardTitle ? t(dashboardTitle.key, dashboardTitle.value) : t('serviceQueues', 'Service queues')}</p>
-          <p className={styles['page-name']}>{title ?? t('home', 'Home')}</p>
-        </div>
-      </div>
-      <div className={styles['right-justified-items']}>
-        <div className={styles['date-and-location']}>
-          <Location size={16} />
-          <span className={styles.value}>{userLocation}</span>
-        </div>
+    <PageHeader
+      dashboardTitle={`${title}`}
+      illustration={<PatientQueueIllustration />}
+      className={styles.header}
+      data-testid="patient-queue-header"
+      utilities={
         <div className={styles.dropdownContainer}>
           {showLocationDropdown && (
             <Dropdown
@@ -97,8 +85,8 @@ const PatientQueueHeader: React.FC<PatientQueueHeaderProps> = ({ title, showLoca
           )}
           {actions}
         </div>
-      </div>
-    </div>
+      }
+    />
   );
 };
