+++ packages/esm-service-queues-app/src/patient-queue-header/patient-queue-header.scss
@@ -13,53 +13,16 @@
   padding-right: 1rem;
 }
 
-.left-justified-items {
+.dropdownContainer {
   display: flex;
-  flex-direction: row;
   align-items: center;
-}
-
-.right-justified-items {
-  @include type.type-style('body-compact-02');
-  color: $text-02;
-  padding-top: 1rem;
-}
-
-.page-name {
-  @include type.type-style('heading-04');
-}
-
-.page-labels {
-  p:first-of-type {
-    margin-bottom: 0.25rem;
-  }
-}
-
-.date-and-location {
-  display: flex;
   justify-content: flex-end;
-  align-items: center;
-}
-
-.value {
-  margin-left: 0.25rem;
-}
-
-.middot {
-  margin: 0 0.5rem;
+  margin: 0.5rem;
 }
 
 .view {
   @include type.type-style('label-01');
 }
-
-.dropdownContainer {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  margin-top: 0.5rem;
-}
-
 .dropdown {
   :global(.cds--list-box__field) {
     width: 14rem;
